\name{Diagnoza-package}
\docType{package}
\title{
Social Diagnosis 2000-2015 data converted to the R format
}
\description{
Disclaimer: I (PBi) have converted the Diagnoza dataset from SPSS to R format to make it easier to use for researchers familiar with R. Diagnoza is an open dataset. I do not hold any rights to it. Use on own risk.

[cited from from http://diagnoza.com/index-en.html]
The Social Diagnosis is a project to support our diagnostical work with detailed data derived from institutional indicators concerning households and the attitudes, mind-sets and behaviours of their members. It is a diagnosis of the conditions and quality of life of the Poles as they report it. We investigate households and their occupants aged 16 and above using two separate questionnaires.

The project takes into account all the significant aspects of the life of individual households and their members, both the economic (income, material wealth, savings and financing), and the not strictly economic (education, medical care, problem-solving, stress, psychological well-being, lifestyle, pathologies, engagement in the arts and cultural events, use of new communication technologies as well as and many others). In this sense the project is interdisciplinary, drawing on the work of the main authors of the Social Monitoring Council (Rada Monitoringu Społecznego) and a team appointed by a Council of experts made up of economists, a demographer, a psychologist, sociologists, an insurance specialist, a health economics expert and statisticians.

As was intended, the Social Diagnosis is based on panel research; we return to the same households every few years, with the first sample being taken in the year 2000. The following took place three years later, and since has been repeated very two years. The reading is always conducted in March to aid the elimination of the seasonal effect. Since 2009 measurement ran into the first half of April due to the size of the sample.

The results of the Social Diagnosis reveal not only the current state of Polish society, but allow us to follow how it has changed over the last ten years. What is more, taking account earlier research on the quality of life in Poland, it even gives us a insight into nearly the entire process of system transformation since 1989.
}
\details{
\tabular{ll}{
Package: \tab Diagnoza\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-11-24\cr
License: \tab see the diagnoza.com website\cr
LazyLoad: \tab yes\cr
}
}
\author{
Przemyslaw Biecek (ICM UW)

Maintainer: Przemyslaw Biecek <Przemyslaw.Biecek@gmail.com>
}
\keyword{package}
\seealso{
}
\examples{
}
